feat: Setup project & Rename packages
