+++ src/main/java/com/studyProjectA/ShoppingMall/dto/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.studyProjectA.ShoppingMall.Dto;
+package com.studyProjectA.ShoppingMall.dto;
 
 import lombok.Data;
